- only save the options that differ from the default values (unless we   explicitly ask to save everything)
- tweak the auto doc routines

+++ Common/Context.h
@@ -123,8 +123,8 @@ public :
   int viewport[4];            // current viewport 
   double vxmin, vxmax, vymin, vymax; // current viewport in real coordinates 
   int light[6];               // status of light 
-  float light_position[6][4]; // light sources positions 
-  float shine;                // specular value 
+  double light_position[6][4]; // light sources positions 
+  double shine;                // specular value 
   int render_mode;            // GMSH_RENDER, GMSH_SELECT, GMSH_FEEDBACK 
   int clip[6];                // status of clip planes 
   double clip_plane[6][4];    // clip planes
